requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==2.0.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jiter==0.5.0
loguru==0.7.2
magic-filter==1.0.12
multidict==6.0.5
openai==1.42.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
yarl==1.9.4
